// types.ts
export interface Car {
  id: number;
  make: string;
  model: string;
  year: number;
  price: number;
  mileage: number;
  transmission: string;
  fuelType: string;
  images: string[];
  description: string;
  features: string[];
  engine: string;
  power: string;
  videoUrl: string;
  marketPriceMorocco: number;
}